diploma_kek/db_setup/requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
filelock==3.6.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
googletrans==4.0.0rc1
greenlet==1.1.2
grpcio==1.46.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.5.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
joblib==1.1.0
libretranslatepy==2.1.1
lxml==4.8.0
Markdown==3.3.7
oauthlib==3.2.0
packaging==21.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
PyYAML==6.0
pyzmq==22.3.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
sacremoses==0.0.53
sentencepiece==0.1.96
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.18.0
translate==3.6.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0